Refactored InitPWM to allow for extra TCC on 64-pin variants of atsamd51
